+++ app/content/login/login.html.js
@@ -2,10 +2,7 @@
 
 /* View Specific JS */
 
-
-/* Import required F1JS modules */
-
-import { Form } from './js/vendors/f1js/form/form.js';
+import Form from './js/vendors/f1js/form.js';
 
 
 F1.deferred.push(function initPage() {
@@ -14,7 +11,7 @@ F1.deferred.push(function initPage() {
   
   const log = console.log;
   const logError = console.error;
-  const loginForm = new Form( { selector: 'form', onlyShowSummary: true });
+  const loginForm = new Form( { selector: 'form' });
 
 
   F1.onSubmit = function( event ) {
@@ -23,7 +20,7 @@ F1.deferred.push(function initPage() {
     const errors = loginForm.getErrors();
     if ( errors.length > 0 ) { 
       event.preventDefault();
-      loginForm.showErrors( errors );
+      loginForm.showErrors({ showSummary: 1 });
       errors[0].focus();
     }
   };