+++ app/themes/salon/doc-foot.html.php
@@ -2,9 +2,7 @@
   <script src="<?=$scriptFileRef?>"></script>
   <?php endforeach; ?>
 
-  <script type="module">
-    <?php include $view->getInlineScriptFile(); ?>
-  </script>
+  <script type="module"><?php include $view->getInlineScriptFile(); ?></script>
 
   <script src="js/main.js"></script>