+++ app/themes/salon/main-menu.html.php
@@ -1,13 +1,13 @@
 <ul class="menu" aria-label="Main Menu">
-  <?php foreach ( $view->menu as $link => $item ): ?>
+  <?php foreach( $view->menu as $link => $item ): ?>
   <?php $isHome = $link==$app->homePage; ?>
   <?php $itemClass =  $isHome ? 'menu-item home' : 'menu-item'; ?>
   <?php $itemLink =  $isHome ? '' : $link; ?>
-  <?php if ( is_array( $item ) ): ?>
+  <?php if( is_array( $item ) ): ?>
   <li class="menu-item has-submenu">
     <a href="<?=$link?>"><?=$item[0]?></a>
     <ul class="submenu">
-    <?php foreach ( $item[1] as $sublink => $subitem ): ?>
+    <?php foreach( $item[1] as $sublink => $subitem ): ?>
       <li><a href="<?=$sublink?>"><?=$subitem?></a></li>
     <?php endforeach; ?>
     </ul>
@@ -17,4 +17,3 @@
   <?php endif; ?>
   <?php endforeach; ?>
 </ul>
-<script src="js/menu.js" deferred></script>